refactor: inline bell-notation conversions as private helpers of bell_inequality_max
The six CG/FC/FP notation conversions lived in their own module but were used
only by bell_inequality_max. Move them into that module as private functions
(_cg_to_fc, _fc_to_cg, _cg_to_fp, _fc_to_fp, _fp_to_cg, _fp_to_fc), drop them
from the public state_opt API, and delete bell_notation_conversions.py.

The conversions do not call each other in code, so this is a mechanical move
plus call-site/import renames. Tests for the conversions keep their coverage by
importing the private names; the error-path tests patch the new private names.
