+++ toqito/state_opt/__init__.py
@@ -11,4 +11,3 @@ from toqito.state_opt.state_exclusion_locc import state_exclusion_locc
 from toqito.state_opt.npa_hierarchy import npa_constraints, bell_npa_constraints
 from toqito.state_opt.symmetric_extension_hierarchy import symmetric_extension_hierarchy
 from toqito.state_opt.bell_inequality_max import bell_inequality_max, bell_inequality_max_qubits
-from toqito.state_opt.bell_notation_conversions import cg_to_fc, cg_to_fp, fc_to_cg, fc_to_fp, fp_to_cg, fp_to_fc
